+++ pkg/traceql/ast_stringer.go
@@ -14,6 +14,10 @@ func (r RootExpr) String() string {
 		s.WriteString(" | ")
 		s.WriteString(r.MetricsPipeline.String())
 	}
+	if r.MetricsSecondStage != nil {
+		s.WriteString(" | ")
+		s.WriteString(r.MetricsSecondStage.String())
+	}
 	if r.Hints != nil {
 		s.WriteString(" ")
 		s.WriteString(r.Hints.String())